Added missing files for lab4 and lab6
Added linux files for lab4. Also added ignoreme.a for lab6 since .gitignore was ignoring .a files previously so it never got merged into the code base. Lastly, added the archives of the original lab files in case I need them again for some reason (they a small so they shouldn't make much of a difference).