Finished lab7
When I first tested it on the Windows machines in the talmage it kept failing. I finally figured out this was because my code didn't deal with the odd case of having a string with only three words. Once I made a separate set of code to deal with this situation, the code ran perfectly and passed all the tests. Hurray!