Reorganized file structure
Since all the projects build off of each other I have decided not
to separate them in the file hierarchy by directories as I have
done in the past as this approach only makes things more
confusing. Instead, all the files will live in the same
directory structure (like that of project01) and will be kept
track of by it. I will simply create a new branch from the
current project so the new project is based off of it. Thus I can
both work straight ahead AND keep track of my projects separately.