cs235/lab03. Made changes to get lab working in MS Windows
Part 2 no longer passes in windows. It seems to struggle when
attempting to close parenthesis. This wasn't a problem
before so I will have to do some digging to figure out what is up.

Had to swap the order of operands for part 3 (postfix to infix)
to pass in Windows.

Part 4 still doesn't work. Actually, it seems to crash the test
driver. This is worse than before.