Rename out61.txt and out62.txt from project 4
Project 5 also uses out61.txt and out62.txt, but the contents are
different. The rename is that we can add project 5 files and still
use them for testing without needing to change our testing scripts.
This will, however, break testing for project 4 on these two files,
but since we have already passed off project 4 this isn't really
a problem.