Add simple optimization
The code previously joined the first relation of a rule with a copy of
itself. With small relations there was no noticable difference, but with
large relations, the time spiked. By making sure the not to join the first
relation with a copy of itself, speed of the program improved by about
440%! On my current home machine it went from around 22 seconds down to
around 5 seconds.